Correct the generation of component with an ending tag, plus refactoring on the cli specs
