+++ package.json
@@ -70,6 +70,7 @@
     "chai": "^3.4.1",
     "eslint": "^1.10.3",
     "eslint-plugin-react": "^3.15.0",
+    "fs-promise": "^0.4.1",
     "mocha": "^2.3.4"
   },
   "dependencies": {
